New file for bundled function
