+++ packages/now-cli/src/util/dev/get-bundled-builders.ts
@@ -1,0 +1,11 @@
+export function getBundledBuilders() {
+  return [
+    '@now/go',
+    '@now/next',
+    '@now/node',
+    '@now/ruby',
+    '@now/python',
+    '@now/static-build',
+    '@now/build-utils',
+  ];
+}
